Clean up a lot of the generated cruft.
It doesn't really help me to know what line of hxDaedalus was translated to produce every single line, especially when I don't know what version of hxDaedalus was used. IDEA wasn't able to handle the large Swing-related files in the tests. I also ran a lot of IDEA inspections on the rest.